NADEModel is fully working. Some minor bugs were fixed

+++ environments/DeterministicMDP.py
@@ -48,18 +48,18 @@ class DeterministicMDP:
 
             # save passed parameters
             self.transition_function = transition_function
-            self.transition = tf.constant(transition_function, dtype=tf.int32)
+            self.transition = tf.constant(transition_function, dtype=tf.int64)
             self.reward_function = reward_function
-            self.rewards = tf.constant(reward_function, dtype=tf.float32)
+            self.rewards = tf.constant(reward_function, dtype=tf.float64)
 
             # Create the current state vector as well as the operation to reset it to the initial state
-            init = tf.constant(self.initial_state, shape=state_tensor_shape, dtype=tf.int32)
-            self.current_states = tf.get_variable("current_state", dtype=tf.int32, initializer=init)
-            self.cum_rewards = tf.get_variable("cum_rewards", state_tensor_shape, dtype=tf.float32, initializer=tf.zeros_initializer)
-            self.eps_rewards = tf.get_variable("eps_rewards", state_tensor_shape, dtype=tf.float32, initializer=tf.zeros_initializer)
+            init = tf.constant(self.initial_state, shape=state_tensor_shape, dtype=tf.int64)
+            self.current_states = tf.get_variable("current_state", dtype=tf.int64, initializer=init)
+            self.cum_rewards = tf.get_variable("cum_rewards", state_tensor_shape, dtype=tf.float64, initializer=tf.zeros_initializer)
+            self.eps_rewards = tf.get_variable("eps_rewards", state_tensor_shape, dtype=tf.float64, initializer=tf.zeros_initializer)
 
             reset_state = tf.assign(self.current_states, init)
-            zero_const = tf.constant(0.0, shape=state_tensor_shape, dtype=tf.float32)
+            zero_const = tf.constant(0.0, shape=state_tensor_shape, dtype=tf.float64)
             reset_cum_rewards = tf.assign(self.cum_rewards, zero_const)
             reset_eps_rewards = tf.assign(self.eps_rewards, zero_const)
