+++ policies/EpsilonGreedyPolicy.py
@@ -24,18 +24,18 @@ class EpsilonGreedyPolicy(Policy):
         best_actions, _ = GreedyPolicy().select_action(q, config)
 
         # sample a random decision vector
-        num_models = tf.shape(q)[0]
+        num_models = tf.cast(tf.shape(q)[0], tf.int64)
 
         # sample some random values
         sample_size = [num_models]
-        random_actions = tf.random_uniform(sample_size, maxval=config['action_space'].get_size(), dtype=tf.int32)
+        random_actions = tf.random_uniform(sample_size, maxval=config['action_space'].get_size(), dtype=tf.int64)
         random_decision_vector = tf.less(tf.random_uniform(sample_size), config['epsilon'])
 
         # let the coin decide about the final actions
         final_actions = tf.where(random_decision_vector, best_actions, random_actions)
 
         # pass back the actions and corresponding q values
-        model_range = tf.range(0, num_models, 1)
+        model_range = tf.range(0, num_models, 1, dtype=tf.int64)
         indices = tf.stack([model_range, final_actions], axis=1)
         q_values = tf.gather_nd(q, indices)
         return final_actions, q_values
